project/requirements/common.txt
dj-database-url==0.4.2
django-configurations==2.0.0
django-debug-toolbar==1.8
python-dotenv==0.6.4
django-extensions==1.7.9
Django==1.11.2
psycopg2==2.7.1
django-bootstrap3==8.2.3
django-simple-captcha==0.5.5
djangorestframework==3.6.2
Markdown==2.6.8
Pygments==2.2.0
coreapi==2.3.1
django-split-settings==0.2.5